requirements.txt
asn1crypto==0.24.0
arrow==0.15.5
bcrypt==3.1.4
BigQuery-Python==1.14.0
boto==2.49.0
cachetools==2.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
cycler==0.10.0
datedelta==1.2
dill==0.2.8.2
distro==1.3.0
dnspython==1.15.0
et-xmlfile==1.0.1
eventlet==0.24.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.4.0
flask-paginate==0.5.1
Flask-SocketIO==3.0.2
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
google-api-core==1.4.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-bigquery==1.6.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.1
google-cloud-core==0.28.1
google-cloud-datastore==1.7.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.2
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.1
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.12.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.0.1
google-cloud-vision==0.29.0
google-compute-engine==2.8.3
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
greenlet==0.4.15
grpc-google-iam-v1==0.11.4
grpcio==1.12.1
gunicorn==19.7.1
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
kiwisolver==1.1.0
lxml==4.2.5
MarkupSafe==1.0
matplotlib==3.1.1
monotonic==1.5
numpy==1.15.1
oauth2client==4.1.2
openpyxl==2.5.7
pandas==0.24.2
pillow==7.2.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.6.1
psutil==5.4.6
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pycryptodome==3.6.6
pydash==4.7.6
pymongo==3.7.1
PyMySQL==0.9.2
pyparsing==2.4.2
python-dateutil==2.7.3
python-engineio==2.2.0
python-http-client==3.1.0
python-socketio==2.0.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
requests-toolbelt==0.8.0
rq==1.5.0
rq-dashboard==0.6.1
rq-scheduler==0.9.1
rq-scheduler-dashboard==0.0.2
rsa==4.0
scipy==1.3.1
seaborn==0.9.0
sendgrid==5.6.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1
XlsxWriter==1.1.0
xlwt==1.3.0
